AmdSevPkg: Base commit copied from OVMF
This commit represents the file copied from OvmfPkg with minor changes
to adjust include paths to be non-local and change the build name.

As more things are pulled over, add their bases to this commit in the
same manner so a diff against what OVMF used to look like is easy to
produce.  Each package copied over needs a new guid, so this changes
as well.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>