AmdSevPkg: create a SEV secret area in the memfd
SEV needs an area to place an injected secret where OVMF can find it
and pass it up as a ConfigurationTable.  This patch implements the
area itself and the SEC part that exposes the area as a structure
containing a GUID so qemu can find it.  The GUID area is 4k aligned
using a rule similar to the reset vector to make for easy searching.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>